# db: restore TrySeekUsingNext across SetOptions batch refresh