The change to batch mutation visibility (#1640) began invalidating the
iterator in SetOptions if the iterator was configured to read through an
indexed batch that had been modified. This prevented the
TrySeekUsingNext optimization from triggering when the pebbleBatch's
iterator was reused after a mutation.

This commit restores this optimization, and allows the batch iterator to
perform TrySeekUsingNext optimizations when the batch view has not been
just refreshed.

Relative to HEAD:
```
name                                                  old time/op    new time/op    delta
MVCCBatchPut_Pebble/valueSize=10/batchSize=1-24         7.52µs ± 4%    7.57µs ± 3%      ~     (p=0.690 n=5+5)
MVCCBatchPut_Pebble/valueSize=10/batchSize=100-24       3.81µs ± 9%    1.39µs ± 2%   -63.52%  (p=0.008 n=5+5)
MVCCBatchPut_Pebble/valueSize=10/batchSize=10000-24     2.83µs ± 6%    1.30µs ± 1%   -54.06%  (p=0.008 n=5+5)
MVCCBatchPut_Pebble/valueSize=10/batchSize=100000-24    2.01µs ± 2%    1.39µs ± 0%   -30.79%  (p=0.008 n=5+5)

name                                                  old speed      new speed      delta
MVCCBatchPut_Pebble/valueSize=10/batchSize=1-24       1.33MB/s ± 4%  1.32MB/s ± 3%      ~     (p=0.587 n=5+5)
MVCCBatchPut_Pebble/valueSize=10/batchSize=100-24     2.63MB/s ±10%  7.19MB/s ± 2%  +173.38%  (p=0.008 n=5+5)
MVCCBatchPut_Pebble/valueSize=10/batchSize=10000-24   3.54MB/s ± 6%  7.69MB/s ± 1%  +117.36%  (p=0.008 n=5+5)
MVCCBatchPut_Pebble/valueSize=10/batchSize=100000-24  4.97MB/s ± 2%  7.17MB/s ± 0%   +44.40%  (p=0.008 n=5+5)
```

Relative to 22.1:
```
name                                                  old time/op    new time/op    delta
MVCCBatchPut_Pebble/valueSize=10/batchSize=1-24         8.31µs ± 3%    7.57µs ± 3%  -8.93%  (p=0.008 n=5+5)
MVCCBatchPut_Pebble/valueSize=10/batchSize=100-24       1.42µs ± 1%    1.39µs ± 2%  -2.14%  (p=0.040 n=5+5)
MVCCBatchPut_Pebble/valueSize=10/batchSize=10000-24     1.35µs ± 1%    1.30µs ± 1%  -3.67%  (p=0.008 n=5+5)
MVCCBatchPut_Pebble/valueSize=10/batchSize=100000-24    1.46µs ± 1%    1.39µs ± 0%  -4.44%  (p=0.008 n=5+5)

name                                                  old speed      new speed      delta
MVCCBatchPut_Pebble/valueSize=10/batchSize=1-24       1.20MB/s ± 4%  1.32MB/s ± 3%  +9.63%  (p=0.008 n=5+5)
MVCCBatchPut_Pebble/valueSize=10/batchSize=100-24     7.04MB/s ± 1%  7.19MB/s ± 2%  +2.19%  (p=0.032 n=5+5)
MVCCBatchPut_Pebble/valueSize=10/batchSize=10000-24   7.41MB/s ± 1%  7.69MB/s ± 1%  +3.78%  (p=0.008 n=5+5)
MVCCBatchPut_Pebble/valueSize=10/batchSize=100000-24  6.86MB/s ± 1%  7.17MB/s ± 0%  +4.64%  (p=0.008 n=5+5)
```

Informs cockroachdb/cockroach#88723.
Informs cockroachdb/cockroach#88819.
Close #1946.